fix the face join function
actually it needed a union-find on the faces to correctly
report them. The face pointer of halfedges is only updated
at the end (as well as the face halfedge pointer and the
removal of faces). I also needed to enrich the face
to store a vector of halfedge to store the holes inside
the face.